Zmienna stabilizacja za bardzo stabilizowała zmienną AverageError w przypadku, gdy AverageError było dużo – poprawiłem to, uzależniając przyrost stabilizacji stosunkiem błędu słowa do AverageError

+++ kurs.hpp
@@ -71,7 +71,7 @@ class Kurs
 		std::vector<double*> new_repetitionsLevels;
 		std::vector<time_t> new_repetitionsTime;
 		std::vector<double> new_repetitionsAverageError;
-		std::vector<unsigned int> new_repetitionsStabilization;
+		std::vector<double> new_repetitionsStabilization;
 		
 		ushort qAllSingleWords;
 		ushort qKnownSingleWords;
